+++ Examples/SwiftUI/Basics/DestinationsSwiftUIExample/Sources/Modules/ColorDetail/ColorDetailState.swift
@@ -11,7 +11,7 @@ import SwiftUI
 import Destinations
 
 @Observable
-final class ColorDetailState: StateModeling {
+final class ColorDetailState: StateModeling, SheetPresenting {
     typealias Destination = ColorDetailView.Destination
     typealias EventType = ColorDetailView.EventType
     typealias InteractorType = Destination.InteractorType
@@ -22,6 +22,8 @@ final class ColorDetailState: StateModeling {
     var colorModel: ColorViewModel?
     var sheetView: ContainerView<AnyView>?
 
+    var sheetPresentation = SheetPresentation()
+    
     init(colorModel: ColorViewModel? = nil) {
         self.colorModel = colorModel
     }
@@ -29,10 +31,31 @@ final class ColorDetailState: StateModeling {
     func handleEvent(_ type: EventType, content: ContentType? = nil) {
         switch type {
             case .colorDetailButton:
-                guard let sheetView else { return }
-                destination?.handleThrowable(closure: { [weak destination] in
-                    try destination?.performAction(for: type, content: .dynamicView(view: sheetView))
+                showSheet(for: type)
+
+        }
+    }
+    
+    private func showSheet(for type: EventType) {
+        buildSheet()
+        guard let sheetView else { return }
+
+        destination?.handleThrowable(closure: { [weak destination] in
+            try destination?.performAction(for: type, content: .dynamicView(view: sheetView))
+        })
+    }
+    
+    private func buildSheet() {
+        sheetView = ContainerView {
+            AnyView(
+                ColorSheetView(colorModel: colorModel, dismissButtonClosure: { @MainActor [weak sheetPresentation] in
+                    sheetPresentation?.dismissSheet()
                 })
+            )
+        }
+
+        sheetPresentation.dismissedClosure = { @MainActor in
+            print("custom sheet dismissed!")
         }
     }